Added packages to separate project files and some clean up.
