+++ app/src/main/AndroidManifest.xml
@@ -35,13 +35,13 @@
         android:theme="@style/Theme.RoboMow"
         tools:targetApi="31">
         <activity
-            android:name=".DeviceListActivity"
+            android:name=".ui.DeviceListActivity"
             android:exported="false" />
         <activity
-            android:name=".JoystickActivity"
+            android:name=".ui.JoystickActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
