+++ app/src/main/res/layout/activity_device_list.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".DeviceListActivity">
+    tools:context=".ui.DeviceListActivity">
 
     <TextView
         android:layout_width="match_parent"
