+++ app/src/main/res/layout/activity_joystick.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".JoystickActivity">
+    tools:context=".ui.JoystickActivity">
 
     <Button
         android:id="@+id/send_button"
